% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_infinite_estimates.R
\name{detect_infinite_estimates}
\alias{detect_infinite_estimates}
\alias{detectInfiniteEstimates}
\title{Detect Infinite Estimates}
\usage{
detect_infinite_estimates(
  x,
  y,
  weights = rep.int(1, nobs),
  start = NULL,
  etastart = NULL,
  mustart = NULL,
  offset = rep.int(0, nobs),
  family = gaussian(),
  control = list(),
  intercept = TRUE,
  singular.ok = TRUE
)
}
\arguments{
\item{x}{\code{x} is a design matrix of dimension \code{n * p}.}

\item{y}{\code{y} is a vector of observations of length \code{n}.}

\item{weights}{an optional vector of \sQuote{prior weights} to be used
    in the fitting process.  Should be \code{NULL} or a numeric vector.}

\item{start}{currently not used.}

\item{etastart}{currently not used.}

\item{mustart}{currently not used.}

\item{offset}{this can be used to specify an \emph{a priori} known
    component to be included in the linear predictor during fitting.
    This should be \code{NULL} or a numeric vector of length equal to
    the number of cases.  One or more \code{\link[stats]{offset}} terms can be
    included in the formula instead or as well, and if more than one is
    specified their sum is used.  See \code{\link[stats]{model.offset}}.}

\item{family}{a description of the error distribution and link
    function to be used in the model.  For \code{glm} this can be a
    character string naming a family function, a family function or the
    result of a call to a family function.  For \code{glm.fit} only the
    third option is supported.  (See \code{\link[stats]{family}} for details of
    family functions.)}

\item{control}{a list of parameters controlling separation
detection. See \code{\link{detect_separation_control}} for
details.}

\item{intercept}{logical. Should an intercept be included in the
    \emph{null} model?}

\item{singular.ok}{logical. If \code{FALSE}, a singular model is an
error.}
}
\description{
Method for \code{\link{glm}} that detects infinite components in
the MLE estimates of generalized linear models with binomial responses.

For all links except the \code{"log"} link, function
\code{detect_infinite_estimates} is a wrapper around
\code{\link{detect_separation}}. For the \code{"log"}
link separated data allocations not necessarily lead to
infinite components in the MLE estimates, therefore another method is used.

\code{\link{detect_infinite_estimates}} for the \code{"log"} link
relies on the linear optimization model developed in
Schwendinger et al. (2021) and for all the other
supported links it relies on the linear programming methods
developed in Konis (2007).
}
\examples{
# The classical example given in Silvapulle (1981) can be utilized
# to show that for the Log-Binomial model there exist data allocations
# which are separated but produce finite estimates.
data("silvapulle1981", package = "detectseparation")

# Since the data is separated the MLE does not exist for the logit link.
glm(y ~ ghqs, data = silvapulle1981, family = binomial(),
    method = "detect_infinite_estimates")

# However, for the log link all components of the MLE are finite.
glm(y ~ ghqs, data = silvapulle1981, family = binomial("log"),
    method = "detect_infinite_estimates")
glm(y ~ ghqs, data = silvapulle1981, family = binomial("log"), start = c(-1, 0))

}
\references{
Silvapulle, M. J. (1981).
On the Existence of Maximum Likelihood Estimators for the Binomial Response Models.
Journal of the Royal Statistical Society. Series B (Methodological), 43(3), 310–313.
\url{http://www.jstor.org/stable/2984941}

Konis K. (2007). *Linear Programming Algorithms for Detecting
Separated Data in Binary Logistic Regression
Models*. DPhil. University of Oxford.
\url{https://ora.ox.ac.uk/objects/uuid:8f9ee0d0-d78e-4101-9ab4-f9cbceed2a2a}

Konis K. (2013). safeBinaryRegression: Safe Binary Regression. R
package version 0.1-3.
\url{https://CRAN.R-project.org/package=safeBinaryRegression}

Kosmidis I. and Firth D. (2021). Jeffreys-prior penalty, finiteness
and shrinkage in binomial-response generalized linear
models. *Biometrika*, **108**, 71–82

Schwendinger, F., Grün, B. & Hornik, K. A comparison of optimization solvers
for log binomial regression including conic programming.
Comput Stat 36, 1721–1754 (2021). \url{https://doi.org/10.1007/s00180-021-01084-5}
}
