% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_infinite_estimates.R
\name{detect_infinite_estimates_control}
\alias{detect_infinite_estimates_control}
\alias{detectInfiniteEstimatesControl}
\title{Auxiliary function for the \code{\link{glm}} interface when
\code{method} is \code{\link{detect_infinite_estimates}}.}
\usage{
detect_infinite_estimates_control(
  solver = "lpsolve",
  tolerance = 1e-04,
  solver_control = list()
)
}
\arguments{
\item{solver}{should the linear program be solved using the
\code{"lpsolve"} (using the \pkg{ROI.plugin.lpsolve} package;
default) or another solver? Alternative solvers are
\code{"glpk"}, \code{"cbc"}, \code{"clp"}, \code{"cplex"},
\code{"ecos"}, \code{"gurobi"}, \code{"scs"},
\code{"symphony"}. If \pkg{ROI.plugin.[solver]} is not
installed then the user will be prompted to install it before
continuing.}

\item{tolerance}{maximum absolute variable value from the linear
program, before separation is declared. Default is
\code{1e-04}.}

\item{solver_control}{a list with additional control parameters for
the \code{"solver"}. This is solver specific, so consult the
corresponding documentation. Default is \code{list()} unless
\code{solver} is \code{"alabama"} when the default is \code{list(start
= rep(0, p))}, where p is the number of parameters.}
}
\value{
A list with the supplied \code{linear_program}, \code{solver},
\code{solver_control}, \code{purpose}, \code{tolerance},
\code{implementation}, and the matched \code{separator} function
(according to the value of \code{implementation}).
}
\description{
Typically only used internally by \code{\link{detect_infinite_estimates}}
but may be used to construct a \code{control} argument.
}
