% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_separation.R, R/zzz_conventions.R
\name{detect_separation_control}
\alias{detect_separation_control}
\alias{detectSeparationControl}
\title{Auxiliary function for the \code{\link{glm}} interface when
\code{method} is \code{\link{detect_separation}}.}
\usage{
detect_separation_control(
  implementation = c("ROI", "lpSolveAPI"),
  solver = "lpsolve",
  linear_program = c("primal", "dual"),
  purpose = c("find", "test"),
  tolerance = sqrt(.Machine$double.eps)
)

detectSeparationControl(
  implementation = c("ROI", "lpSolveAPI"),
  solver = "lpsolve",
  linear_program = c("primal", "dual"),
  purpose = c("find", "test"),
  tolerance = sqrt(.Machine$double.eps)
)
}
\arguments{
\item{implementation}{should the implementation using \code{ROI} or
the implementation usinf \code{lpSolveAPI} be used? Default is
\code{ROI}.}

\item{solver}{should the linear program be solved using the
\code{"lpsolve"} (using the **ROI.plugin.lpsolve** package;
default) or the \code{"glpk"} (using the **ROI.plugin.glpk**
package) solver?}

\item{linear_program}{should \code{\link{detect_separation}} solve
the \code{"primal"} or \code{"dual"} linear program for
separation detection? Only relevant if \code{implementation =
"lpSolveAPI"}.}

\item{purpose}{should \code{\link{detect_separation}} simply
\code{"test"} for separation or also \code{"find"} which
parameters are infinite? Only relevant if \code{implementation
= "lpSolveAPI"}.}

\item{tolerance}{maximum absolute variable value from the linear
program, before separation is declared.}
}
\description{
Typically only used internally by \code{\link{detect_separation}}
but may be used to construct a \code{control} argument.
}
