% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{silvapulle1981}
\alias{silvapulle1981}
\title{Separation Example Presented in Silvapulle (1981)}
\format{
A data frame with 35 rows and 2 variables:
\itemize{
\item \code{y}: a factor with the levels \code{case} and \code{none-case}, giving the
outcome of a standardized psychiatric interview
\item \code{ghqs}: an integer giving the general health questionnaire score.
}
}
\usage{
silvapulle1981
}
\description{
Separation example presented in Silvapulle (1981).
}
\references{
Silvapulle, M. J. (1981).
On the Existence of Maximum Likelihood Estimators for the Binomial Response Models.
Journal of the Royal Statistical Society. Series B (Methodological), 43(3), 310–313.
\url{https://www.jstor.org/stable/2984941}
}
\seealso{
\code{\link[=detect_infinite_estimates]{detect_infinite_estimates()}}
}
\keyword{datasets}
