% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detectseparation-package.R
\docType{package}
\name{detectseparation}
\alias{detectseparation}
\title{detectseparation: Detect or check for separation and infinite maximum likelihood esitmates in generalized linear models with categorical responses}
\description{
[SHORT DESCRIPTION HERE]
}
