% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detectseparation-package.R
\docType{package}
\name{detectseparation}
\alias{detectseparation}
\title{detectseparation: Methods for Detecting and Checking for Separation
and Infinite Maximum Likelihood Estimates}
\description{
detectseparation: Methods for Detecting and Checking for Separation
and Infinite Maximum Likelihood Estimates
}
\seealso{
\code{\link{detect_separation}}, \code{\link{check_infinite_estimates}}
}
